% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing.R
\name{check_missing}
\alias{check_missing}
\title{Figure out the variables having missing values}
\usage{
check_missing(df, spaces = FALSE)
}
\arguments{
\item{The}{input to the function is a data frame which is mandatory. The optional parameter is spaces. The code by default takes spaces as FALSE but you can override it to convert spaces to NAs and return the missing value results.}
}
\value{
Returns a vector of column names in the data frame which have missing values.
If spaces parameter is over-ridden it returns the variables with single spaces and also treats and returns them as NAs as well.
}
\description{
\code{check_missing} identifies the variables in the data frame with missing values and
returns their variable names.
}
\examples{
zero_var_index <- zero_var(df)
df_without_zerovariance <- df[,-zero_var_index]
}
