% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_var.R
\name{zero_var}
\alias{zero_var}
\title{Figure out the variables having zero variance}
\usage{
zero_var(df)
}
\arguments{
\item{The}{input to the function is a data frame.}
}
\value{
Returns a vector of column names in the data frame which have zero variance
along with their index.
}
\description{
\code{zero_var} identifies the variables in the data frame with zero variance and
returns their index values. The code has been by Gavin Simpson and taken from
stackoverflow.
}
\examples{
zero_var_index <- zero_var(df)
df_without_zerovariance <- df[,-zero_var_index]
}
